working on regularactivity

+++ app/src/main/java/com/example/harpocrates/account/InexistantUserException.java
@@ -1,7 +1,7 @@
 package com.example.harpocrates.account;
 
 public class InexistantUserException extends Exception {
-    public InexistantUserException( int id ) {
+    public InexistantUserException( long id ) {
         super( "the user_id " + id + " does not exist in our database" );
     }
 }
